Refactor Badge Icon
