+++ packages/medulas-react-components/src/components/Badge/index.tsx
@@ -1,0 +1,50 @@
+import { Badge, makeStyles, PropTypes, Theme } from '@material-ui/core';
+import React from 'react';
+import Img from '../Image';
+import CheckIcon from './assets/check.svg';
+
+interface Props extends React.ImgHTMLAttributes<HTMLImageElement> {
+  readonly variant: 'dot' | 'check';
+  readonly invisible?: boolean;
+  readonly color?: PropTypes.Color | 'error';
+  readonly children: React.ReactNode;
+}
+
+const useStyles = makeStyles((theme: Theme) => ({
+  check: {
+    width: '25px',
+    height: '25px',
+    top: '-8px',
+    right: '-8px',
+  },
+  dot: {
+    width: '7px',
+    height: '7px',
+    top: '-7px',
+    right: '-7px',
+  },
+}));
+
+const BadgeIcon = ({ variant, invisible = false, color = 'primary', children }: Props): JSX.Element => {
+  const classes = useStyles();
+
+  const badgeClasses = { badge: variant === 'check' ? classes.check : classes.dot };
+  const badgeContent = variant === 'check' ? <Img src={CheckIcon} alt="Badge Icon" /> : '';
+
+  const badgeVariant = variant === 'check' ? 'standard' : 'dot';
+  const badgeColor = variant === 'check' ? 'default' : color;
+
+  return (
+    <Badge
+      badgeContent={badgeContent}
+      classes={badgeClasses}
+      invisible={invisible}
+      variant={badgeVariant}
+      color={badgeColor}
+    >
+      {children}
+    </Badge>
+  );
+};
+
+export default BadgeIcon;
